Better align guard/multiif/case
This is tricky since they are several variants of the same concept
but do behave differently

Cases can be guarded or unguarded. Unguarded path needs to holistically
tabstop with all alternatives and guards

multiif pads differently

function guards are the simplest, but different from the other to.

Additionally, i am not even handling `case () of _ |`, as it is even
more complex and special.

For now opt into caseAlign or multiIf align makes it all align. May
revisit later.